refactor: use context for tools actions
