+++ x-pack/platform/plugins/shared/onechat/public/application/hooks/tools/use_tools.ts
@@ -9,15 +9,13 @@ import { ToolType } from '@kbn/onechat-common';
 import { EsqlToolDefinitionWithSchema } from '@kbn/onechat-common/tools/esql';
 import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
+import { useToolsPreferences } from '../../context/tools_preferences_provider';
 import { queryKeys } from '../../query_keys';
 import { useOnechatServices } from '../use_onechat_service';
 
-export interface UseOnechatToolsOptions {
-  includeSystemTools?: boolean;
-}
-
-export const useOnechatTools = ({ includeSystemTools = true }: UseOnechatToolsOptions = {}) => {
+export const useOnechatTools = () => {
   const { toolsService } = useOnechatServices();
+  const { includeSystemTools } = useToolsPreferences();
 
   const { data, isLoading, error } = useQuery({
     queryKey: queryKeys.tools.all,
