+++ x-pack/platform/plugins/shared/onechat/public/application/pages/tools.tsx
@@ -11,12 +11,15 @@ import { ToolsPreferencesProvider } from '../context/tools_preferences_provider'
 import { useBreadcrumb } from '../hooks/use_breadcrumbs';
 import { appPaths } from '../utils/app_paths';
 import { labels } from '../utils/i18n';
+import { ToolsProvider } from '../context/tools_provider';
 
 export const OnechatToolsPage = () => {
   useBreadcrumb([{ text: labels.tools.title, path: appPaths.tools.list }]);
   return (
     <ToolsPreferencesProvider>
-      <OnechatTools />
+      <ToolsProvider>
+        <OnechatTools />
+      </ToolsProvider>
     </ToolsPreferencesProvider>
   );
 };
